pyproject.toml
[tool.poetry]
name = "deepl-haystack"
version = "0.1.0"
description = "Haystack integration with DeepL translation services provider."
authors = ["Albert Iribarne <iribarne@dribia.com>"]
maintainers = ["Albert Iribarne <iribarne@dribia.com>", "Dribia Data Research <code@dribia.com>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/dribia/deepl-haystack"
homepage = "https://github.com/dribia/deepl-haystack"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["haystack", "deepl", "translation", "nlp"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/dribia/deepl-haystack/issues"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
deepl = "^1.18.0"
haystack-ai = "^2.3.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.1"
pytest-mock = "^3.14.0"
ruff = "0.5.4"
mypy = "1.11.0"
pytest-cov = "^5.0.0"

[tool.poetry.group.lint.dependencies]
ruff = "^0.5.4"
mypy = "^1.11.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[tool.ruff]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "D",  # pydocstyle
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "C901",  # function is too complex
    "D206",  # indent with spaces, not recommended with formatter
    "W191",  # tab-indentation, not recommended with formatter
    "D203",  # no-blank-line-before-class
    "D213",  # multi-line-summary-second-line
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]

[tool.ruff.lint.isort]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
plugins = []
strict_optional = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "haystack",
    "haystack.utils"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--verbose"
filterwarnings = 'error'
xfail_strict = true
testpaths = ["tests"]
markers = []

[tool.coverage.run]
source = ["deepl_haystack", "tests"]
branch = true

[tool.coverage.report]
precision = 2
exclude_lines = [
    'pragma: no cover',
    'raise NotImplementedError',
    'if TYPE_CHECKING:',
    '@overload',
]

[tool.coverage.xml]
output = "coverage.xml"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
